# lib/api-zod/src/generated/types/architectureSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ArchitectureSession {
  id: number;
  buildingType: string;
  buildingSubtype: string;
  area: number;
  floors: number;
  additionalRequirements?: string;
  generatedPlan: string;
  conversationId: number;
  createdAt: Date;
}