# lib/api-zod/src/generated/types/createArchitectureSessionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateArchitectureSessionBodyBuildingType } from "./createArchitectureSessionBodyBuildingType";

export interface CreateArchitectureSessionBody {
  buildingType: CreateArchitectureSessionBodyBuildingType;
  buildingSubtype: string;
  area: number;
  floors: number;
  additionalRequirements?: string;
}